move usemakefilters to helpers so it can be reused by multiple components
