+++ moped-editor/src/components/GridTable/helpers.js
@@ -1,5 +1,6 @@
 import { OPERATORS_WITHOUT_SEARCH_VALUES } from "src/views/projects/projectsListView/ProjectsListViewFiltersConf";
 import { AUTOCOMPLETE_OPERATORS } from "src/views/projects/projectsListView/ProjectsListViewFiltersConf";
+import { useMemo } from "react";
 
 /**
  * Generates a copy of an empty filter
@@ -136,3 +137,28 @@ export const getDefaultOperator = (filterConfigForField) => {
 
   return isDefaultOperator ? defaultOperator : fallbackOperator;
 };
+
+/**
+ * if filter exists in url, get the values and try to parse them
+ * Used to initialize filter state
+ * @return Object
+ */
+export const useMakeFilterState = (
+  searchParams,
+  advancedSearchFilterParamName
+) =>
+  useMemo(() => {
+    if (Array.from(searchParams).length > 0) {
+      const filterSearchParams = searchParams.get(
+        advancedSearchFilterParamName
+      );
+      if (filterSearchParams === null) return [];
+
+      try {
+        return JSON.parse(filterSearchParams);
+      } catch {
+        return [];
+      }
+    }
+    return [];
+  }, [searchParams, advancedSearchFilterParamName]);
